de-intern-2024-project/requirements.txt
# AWS SDK and Tools
boto3>=1.34.0
botocore>=1.34.0
awscli>=1.32.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
alembic>=1.13.0

# AWS Glue / PySpark (for local development)
pyspark>=3.5.0

# Data Validation
great-expectations>=0.18.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.0
pyyaml>=6.0.0
tabulate>=0.9.0

# Logging and Monitoring
python-json-logger>=2.0.7

# Date and Time
python-dateutil>=2.8.2
pytz>=2023.3
